Pipeline infrastructure

+++ scripts/add_title_geo_and_CNR_to_first_line.py
@@ -1,29 +1,42 @@
 import os
 
+import re
+
 import pandas as pd
 
+from scripts.Constants import METADATA
 from scripts.task import Task
+from scripts.util import get_target_case
+
 
 
 class AddTitleGeoCNRToFirstLine(Task):
 
-    def __init__(self, input_dir: str):
+
+    def __init__(self, input_dir: str, pattern: str, use_cnr: bool = False):
         super().__init__()
+        self.use_cnr = use_cnr
         self.input_dir = input_dir
+        try:
+            self.pattern = re.compile(pattern)
+        except re.error:
+            raise Exception({"message": "given pattern invalid"})
 
     def _execute(self):
-        assert isinstance(self.shared_resource, pd.DataFrame)
+        metadata = self.shared_resources[METADATA]
+        assert isinstance(metadata, pd.DataFrame)
         for filename in os.listdir(self.input_dir):
             if filename.endswith(".txt"):
                 judgement = os.path.join(self.input_dir, filename)
-                case_number = filename.split('_')[0]
-                case = self.shared_resource[self.shared_resource['case_no'] == case_number]
-                title = self._get_case_title(case)
-                with open(judgement, "r+") as input_file:
-                    judgement = input_file.read()
-                    input_file.seek(0)
-                    input_file.write("{} ||| {}".format(title, judgement))
-                    input_file.truncate()
+                case = get_target_case(filename, metadata, self.use_cnr, self.pattern)
+                if case is not None and not case.empty:
+                    title = self._get_case_title(case)
+                    with open(judgement, "r+") as input_file:
+                        judgement = input_file.read()
+                        input_file.seek(0)
+                        input_file.write("{} ||| {}".format(title, judgement))
+                        input_file.truncate()
+
 
     @staticmethod
     def _get_case_title(case):
@@ -35,4 +48,6 @@ class AddTitleGeoCNRToFirstLine(Task):
         year = case['reg_year'].item()
         status = 'Pending' if case['disp_nature'].item() == 0 else "Disposed"
 
-        return "{} - {} vs {} - {} - {} - {} - {}".format(CNR, petitioner, respondent, court, case_number, year, status)
+        return "{} - {} vs {} - {}".format(CNR, petitioner, respondent, court)
+        //return "{} - {} vs {} - {} - {} - {} - {}".format(CNR, petitioner, respondent, court, case_number, year, status)
+
